Factor out a Message parent of LogEntry.
We have several messages which are not logged. Factor Message
out of LogEntry so that we can reuse the serialization code
for those.

Currently descriptors and entry classes are 1:1. It may simplify
matters to merge the two abstractions, so there is just one
factory for Entry types. Need to think some more about that.